third_party/WebKit/LayoutTests/fast/dom/HTMLDocument/document-all.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var htmlallcollection = document.all;
shouldBe('htmlallcollection.toString()', "'[object HTMLAllCollection]'");
</script>
</body>
</html>
