third_party/blink/web_tests/platform/generic/fast/dom/HTMLDocument/document-all-expected.txt

PASS htmlallcollection.toString() is '[object HTMLAllCollection]'
PASS successfullyParsed is true

TEST COMPLETE
